% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{print_results}
\alias{print_results}
\title{Pretty-print the results of analyze_lmm.}
\usage{
print_results(results)
}
\arguments{
\item{results}{A named list returned by \code{\link{analyze_lmm}}.}
}
\value{
Invisibly returns \code{results} (called for its side effects).
}
\description{
Pretty-print the results of analyze_lmm.
}
